Revert "minor UI fixes (#9)"
This reverts commit eb89124e92d9bee22f351e1c438b5fd152410be0.
