+++ web/main.js
@@ -1,6 +1,5 @@
 import { streamGemini } from './gemini-api.js';
 import { generateMermaid } from './generateMermaid.js';
-import { detectSpeakingTestIntent, initiateSpeakingTest, isSpeakingTestActive } from './speakingTest.js';
 
 // DOM elements
 const form = document.querySelector('.input-form');
@@ -19,11 +18,11 @@ let attachedImageData = null;
 let conversationHistory = [
   {
     role: 'user',
-    parts: [{ text: "You are Sahayak, a helpful AI assistant with various tools. When users want to practice speaking, take a speaking test, or improve their pronunciation, you should automatically initiate the speaking test tool. Please introduce yourself as Sahayak and be friendly and helpful." }]
+    parts: [{ text: "You are Sahayak, a helpful AI assistant. Please introduce yourself as Sahayak and be friendly and helpful." }]
   },
   {
     role: 'model',
-    parts: [{ text: "Hello! I'm Sahayak, your AI assistant. I'm here to help you with any questions or tasks you have. I can also help you practice speaking and pronunciation - just let me know if you'd like to take a speaking test! How can I assist you today?" }]
+    parts: [{ text: "Hello! I'm Sahayak, your AI assistant. I'm here to help you with any questions or tasks you have. How can I assist you today?" }]
   }
 ];
 
@@ -82,18 +81,6 @@ document.addEventListener('DOMContentLoaded', () => {
 
     if (!message && !attachedImageData) {
       return;
-    }
-
-    // Check for speaking test intent first
-    if (message) {
-      const isSpeakingIntent = await detectSpeakingTestIntent(message, conversationHistory);
-      if (isSpeakingIntent) {
-        addUserMessage(message);
-        messageInput.value = '';
-        await initiateSpeakingTest(addAssistantMessage, scrollToBottom, conversationHistory);
-        messageInput.focus();
-        return;
-      }
     }
 
     // Disable form
@@ -260,8 +247,7 @@ function renderMermaidDiagram(code, container) {
     window.mermaid.init(undefined, mermaidDiv);
   }
 }
-
-// --- Helper Functions ---
+// Helper functions
 function addUserMessage(text, imageUrl = null) {
   const messageDiv = document.createElement('div');
   messageDiv.className = 'message user-message';
@@ -319,3 +305,113 @@ function escapeHtml(text) {
   div.textContent = text;
   return div.innerHTML;
 }
+
+// --- Reading Tutor Logic ---
+
+const showTutorBtn = document.getElementById('show-tutor-btn');
+const readingTutorContainer = document.querySelector('.reading-tutor-container');
+const startRecordBtn = document.getElementById('start-record-btn');
+const stopRecordBtn = document.getElementById('stop-record-btn');
+const readingText = document.getElementById('reading-text');
+const readingResultsDiv = document.getElementById('reading-results');
+const readingResultsContent = readingResultsDiv.querySelector('.message-content');
+
+let mediaRecorder;
+let audioChunks = [];
+
+const paragraphs = [
+    "The sun dipped below the horizon, painting the sky in shades of orange and pink. A gentle breeze rustled the leaves in the trees, creating a soft, whispering sound.",
+    "Technology has advanced at an incredible pace over the last few decades. From the first computers to the powerful devices in our pockets, the change has been monumental.",
+    "A balanced diet and regular exercise are crucial for maintaining good health. It is important to consume a variety of nutrients and stay active to keep your body and mind in top shape.",
+    "The old library was a quiet sanctuary, filled with the scent of aged paper and leather-bound books. Every shelf held stories waiting to be discovered by an eager reader."
+];
+
+showTutorBtn.addEventListener('click', () => {
+    const randomIndex = Math.floor(Math.random() * paragraphs.length);
+    readingText.innerText = paragraphs[randomIndex];
+    readingResultsDiv.style.display = 'none';
+    readingResultsContent.innerHTML = '';
+    readingTutorContainer.style.display = 'block';
+    startRecordBtn.disabled = false;
+    startRecordBtn.innerText = 'Start Recording';
+    stopRecordBtn.disabled = true;
+});
+
+startRecordBtn.addEventListener('click', async () => {
+  if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
+    alert('Your browser does not support audio recording.');
+    return;
+  }
+
+  try {
+    const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
+    
+    const options = { mimeType: 'audio/webm;codecs=opus' };
+    if (!MediaRecorder.isTypeSupported(options.mimeType)) {
+        console.warn(`${options.mimeType} is not supported, using browser default.`);
+        mediaRecorder = new MediaRecorder(stream);
+    } else {
+        mediaRecorder = new MediaRecorder(stream, options);
+    }
+
+    mediaRecorder.ondataavailable = (event) => {
+      audioChunks.push(event.data);
+    };
+
+    mediaRecorder.onstop = async () => {
+      const audioBlob = new Blob(audioChunks, { type: options.mimeType });
+      audioChunks = [];
+
+      if (audioBlob.size === 0) {
+        alert("Recording was empty. Please try recording for a longer duration.");
+        stopRecordBtn.disabled = true;
+        startRecordBtn.disabled = false;
+        startRecordBtn.innerText = 'Start Recording';
+        return;
+      }
+
+      readingResultsContent.innerHTML = 'Analyzing your reading... Please wait.';
+      readingResultsDiv.style.display = 'block';
+
+      const formData = new FormData();
+      formData.append('audio_file', audioBlob, 'recording.webm');
+      formData.append('original_text', readingText.innerText);
+
+      try {
+        const response = await fetch('/api/analyze_reading', {
+          method: 'POST',
+          body: formData,
+        });
+
+        const result = await response.json();
+
+        if (result.error) {
+          readingResultsContent.innerText = `Error: ${result.error}`;
+        } else {
+          const md = new markdownit();
+          readingResultsContent.innerHTML = md.render(result.analysis);
+        }
+      } catch (error) {
+        readingResultsContent.innerText = `An unexpected error occurred: ${error.message}`;
+      }
+
+      stopRecordBtn.disabled = true;
+      startRecordBtn.disabled = false;
+      startRecordBtn.innerText = 'Start Recording';
+    };
+
+    mediaRecorder.start();
+    startRecordBtn.disabled = true;
+    startRecordBtn.innerText = 'Recording...';
+    stopRecordBtn.disabled = false;
+  } catch (error) {
+    alert(`Error starting recording: ${error.message}`);
+    startRecordBtn.disabled = false;
+  }
+});
+
+stopRecordBtn.addEventListener('click', () => {
+  if (mediaRecorder && mediaRecorder.state === 'recording') {
+    mediaRecorder.stop();
+  }
+});
