+++ web/meeting_dashboard.html
@@ -20,34 +20,6 @@
             background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
             color: white;
             border-radius: 10px;
-            position: relative;
-        }
-
-        .back-button {
-            position: absolute;
-            left: 20px;
-            top: 50%;
-            transform: translateY(-50%);
-            background: rgba(255, 255, 255, 0.2);
-            color: white;
-            border: 2px solid rgba(255, 255, 255, 0.3);
-            padding: 10px 15px;
-            border-radius: 25px;
-            text-decoration: none;
-            font-size: 14px;
-            font-weight: 500;
-            transition: all 0.3s ease;
-            display: flex;
-            align-items: center;
-            gap: 8px;
-        }
-
-        .back-button:hover {
-            background: rgba(255, 255, 255, 0.3);
-            border-color: rgba(255, 255, 255, 0.5);
-            transform: translateY(-50%) scale(1.05);
-            color: white;
-            text-decoration: none;
         }
 
         .dashboard-grid {
@@ -298,19 +270,6 @@
             .btn {
                 width: 100%;
             }
-
-            .back-button {
-                position: static;
-                transform: none;
-                margin-bottom: 15px;
-                align-self: flex-start;
-            }
-
-            .dashboard-header {
-                display: flex;
-                flex-direction: column;
-                align-items: center;
-            }
         }
 
         /* Markdown Content Styling */
@@ -435,10 +394,6 @@
 <body>
     <div class="meeting-dashboard">
         <div class="dashboard-header">
-            <a href="/" class="back-button">
-                <span>←</span>
-                <span>Back</span>
-            </a>
             <h1>🏫 Sahayak - Parent-Teacher Meeting Assistant</h1>
             <p>AI-Powered Talking Points Generator for Effective Parent Conversations</p>
         </div>
